Final changes!

+++ compiledcode.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 int main() {
-	float a = 0;
+	int a = 0;
 	std::cout << "Hello, world!" << std::endl;
 	return 0;
 }
